fix: serialize SMART attrsJson for macOS IPC, fix WCHAR->UTF8 in JSON, fix health type
- main_mac.cpp: serialize SMART attributes to JSON attrsJson, register attrCount/attrsJson in schema
- IPCData.h: add attrCount field to PhysDiskSlot
- IPCSystemInfoMapper.cs: fix health read from UInt8 to Float32 (matches macOS schema)

+++ AvaloniaUI/Services/IPCServices/IPCSystemInfoMapper.cs
@@ -193,7 +193,7 @@ public static class IPCSystemInfoMapper
                     var capacity = reader.ReadUInt64($"phys/{idx}/capacity") ?? 0;
                     var iface = reader.ReadWString($"phys/{idx}/interface") ?? "";
                     var temp = reader.ReadFloat64($"phys/{idx}/temperature") ?? 0;
-                    var health = reader.ReadUInt8($"phys/{idx}/health") ?? 0;
+                    var health = (byte)(reader.ReadFloat32($"phys/{idx}/health") ?? 0f);
                     var supported = reader.ReadBool($"phys/{idx}/smartSupported") ?? false;
 
                     // Read logical drive letters
